Used in_toto.util's functions, edited util.py to include the bits parameter, edited tests
